+++ components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/DefaultTopSitesStorage.kt
@@ -4,17 +4,24 @@
 
 package mozilla.components.feature.top.sites
 
+import android.content.Context
+import androidx.annotation.VisibleForTesting
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import mozilla.components.browser.state.search.RegionState
 import mozilla.components.browser.storage.sync.PlacesHistoryStorage
 import mozilla.components.concept.storage.FrecencyThresholdOption
+import mozilla.components.feature.top.sites.TopSite.Type.DEFAULT
 import mozilla.components.feature.top.sites.TopSite.Type.FRECENT
+import mozilla.components.feature.top.sites.db.DefaultSite
 import mozilla.components.feature.top.sites.ext.hasUrl
 import mozilla.components.feature.top.sites.ext.toTopSite
 import mozilla.components.feature.top.sites.facts.emitTopSitesCountFact
 import mozilla.components.support.base.observer.Observable
 import mozilla.components.support.base.observer.ObserverRegistry
+import mozilla.components.support.locale.LocaleManager
+import java.util.Locale
 import kotlin.coroutines.CoroutineContext
 
 /**
@@ -23,27 +30,160 @@ import kotlin.coroutines.CoroutineContext
  * @param pinnedSitesStorage An instance of [PinnedSiteStorage], used for storing pinned sites.
  * @param historyStorage An instance of [PlacesHistoryStorage], used for retrieving top frecent
  * sites from history.
- * @param defaultTopSites A list containing a title to url pair of default top sites to be added
- * to the [PinnedSiteStorage].
+ * @param region The region of the user.
+ * @param isDefaultSiteAdded Whether or not the default sites were already added from Fenix.
  */
+@Suppress("LongParameterList")
 class DefaultTopSitesStorage(
+    context: Context,
     private val pinnedSitesStorage: PinnedSiteStorage,
     private val historyStorage: PlacesHistoryStorage,
-    private val defaultTopSites: List<Pair<String, String>> = listOf(),
+    private val region: RegionState?,
+    private val isDefaultSiteAdded: Boolean,
     coroutineContext: CoroutineContext = Dispatchers.IO
 ) : TopSitesStorage, Observable<TopSitesStorage.Observer> by ObserverRegistry() {
 
+    private val sharedPref = context.getSharedPreferences(PREFERENCE_NAME, Context.MODE_PRIVATE)
     private var scope = CoroutineScope(coroutineContext)
 
     // Cache of the last retrieved top sites
     var cachedTopSites = listOf<TopSite>()
 
     init {
-        if (defaultTopSites.isNotEmpty()) {
+        if (sharedPref.getInt(MOZAC_LOCAL_DEFAULT_TOP_SITES_VERSION, 0) !=
+            sharedPref.getInt(MOZAC_DEFAULT_TOP_SITES_VERSION, 1)
+        ) {
             scope.launch {
-                pinnedSitesStorage.addAllPinnedSites(defaultTopSites, isDefault = true)
+                initializeDefaultTopSites(context)
+
+                if (!isDefaultSiteAdded) {
+                    // Add default top sites for new installs into the pinned sites database table
+                    // based on the user's region and language.
+
+                    // If the current region is not CN, RU, TR, KZ or BY, default to a generic XX
+                    // region.
+                    val countryCode =
+                        if (region?.current != null && listOf(
+                                CN_REGION,
+                                RU_REGION,
+                                TR_REGION,
+                                KZ_REGION,
+                                BY_REGION
+                            ).contains(region.current)
+                        ) region.current else XX_REGION
+
+                    val locale = LocaleManager.getCurrentLocale(context)
+                        ?: LocaleManager.getSystemDefault()
+                    val language =
+                        if (locale.language == Locale("en").language) EN_LANGUAGE else XX_LANGUAGE
+
+                    val defaultTopSites =
+                        pinnedSitesStorage.getDefaultSites(countryCode, language)
+                            .map { entity -> Pair(entity.title, entity.url) }
+                    pinnedSitesStorage.addAllPinnedSites(defaultTopSites, isDefault = true)
+                } else if (listOf(CN_REGION, RU_REGION, TR_REGION, KZ_REGION, BY_REGION).contains(
+                        region?.current
+                    )
+                ) {
+                    // Remove existing Google default top sites for CN, RU, TR, KZ, BY regions.
+                    pinnedSitesStorage.getPinnedSites()
+                        .find { it.type == DEFAULT && it.url == GOOGLE_URL }?.let {
+                            removeTopSite(it)
+                        }
+                }
             }
+
+            sharedPref.edit().putInt(MOZAC_LOCAL_DEFAULT_TOP_SITES_VERSION, 1).apply()
         }
+    }
+
+    /**
+     * Adds all the default top sites into the default top sites database table.
+     */
+    @Suppress("LongMethod")
+    private suspend fun initializeDefaultTopSites(context: Context) {
+        pinnedSitesStorage.addAllDefaultSites(
+            listOf(
+                DefaultSite(
+                    CN_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_baidu),
+                    BAIDU_URL
+                ),
+                DefaultSite(
+                    CN_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_jd),
+                    JD_URL
+                ),
+                DefaultSite(
+                    RU_REGION,
+                    EN_LANGUAGE,
+                    context.getString(R.string.pocket_pinned_top_articles),
+                    POCKET_TRENDING_URL
+                ),
+                DefaultSite(
+                    RU_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_wikipedia),
+                    WIKIPEDIA_URL
+                ),
+                DefaultSite(
+                    TR_REGION,
+                    EN_LANGUAGE,
+                    context.getString(R.string.pocket_pinned_top_articles),
+                    POCKET_TRENDING_URL
+                ),
+                DefaultSite(
+                    TR_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_wikipedia),
+                    WIKIPEDIA_URL
+                ),
+                DefaultSite(
+                    KZ_REGION,
+                    EN_LANGUAGE,
+                    context.getString(R.string.pocket_pinned_top_articles),
+                    POCKET_TRENDING_URL
+                ),
+                DefaultSite(
+                    KZ_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_wikipedia),
+                    WIKIPEDIA_URL
+                ),
+                DefaultSite(
+                    BY_REGION,
+                    EN_LANGUAGE,
+                    context.getString(R.string.pocket_pinned_top_articles),
+                    POCKET_TRENDING_URL
+                ),
+                DefaultSite(
+                    BY_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_wikipedia),
+                    WIKIPEDIA_URL
+                ),
+                DefaultSite(
+                    XX_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_google),
+                    GOOGLE_URL
+                ),
+                DefaultSite(
+                    XX_REGION,
+                    EN_LANGUAGE,
+                    context.getString(R.string.pocket_pinned_top_articles),
+                    POCKET_TRENDING_URL
+                ),
+                DefaultSite(
+                    XX_REGION,
+                    XX_LANGUAGE,
+                    context.getString(R.string.default_top_site_wikipedia),
+                    WIKIPEDIA_URL
+                )
+            )
+        )
     }
 
     override fun addTopSite(title: String, url: String, isDefault: Boolean) {
@@ -102,5 +242,38 @@ class DefaultTopSitesStorage(
         cachedTopSites = topSites
 
         return topSites
+    }
+
+    companion object {
+        internal const val PREFERENCE_NAME = "mozac_feature_top_sites"
+
+        // The user's local default top sites version.
+        @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+        internal const val MOZAC_LOCAL_DEFAULT_TOP_SITES_VERSION =
+            "MOZAC_LOCAL_DEFAULT_TOP_SITES_VERSION"
+        // This is a MOZAC managed versioning number to upgrade default top sites in the future.
+        private const val MOZAC_DEFAULT_TOP_SITES_VERSION =
+            "MOZAC_DEFAULT_TOP_SITES_VERSION"
+
+        internal const val CN_REGION = "CN"
+        internal const val RU_REGION = "RU"
+        internal const val TR_REGION = "TR"
+        internal const val KZ_REGION = "KZ"
+        internal const val BY_REGION = "BY"
+        // Catch all for other regions that does not fall into CN, RU, TR, KZ, or BY.
+        internal const val XX_REGION = "XX"
+
+        internal const val EN_LANGUAGE = "en"
+        // Catch all for other languages that does not fall into "en".
+        internal const val XX_LANGUAGE = "XX"
+
+        internal const val POCKET_TRENDING_URL = "https://getpocket.com/fenix-top-articles"
+        internal const val WIKIPEDIA_URL = "https://www.wikipedia.org/"
+        internal const val GOOGLE_URL = "https://www.google.com/"
+        internal const val BAIDU_URL = "https://m.baidu.com/?from=1000969a"
+        internal const val JD_URL = "https://union-click.jd.com/jdc" +
+            "?e=&p=AyIGZRprFDJWWA1FBCVbV0IUWVALHFRBEwQAQB1AWQkFVUVXfFkAF14lRFRbJXstVWR3WQ1rJ08AZnhS" +
+            "HDJBYh4LZR9eEAMUBlccWCUBEQZRGFoXCxc3ZRteJUl8BmUZWhQ" +
+            "AEwdRGF0cMhIAVB5ZFAETBVAaXRwyFQdcKydLSUpaCEtYFAIXN2UrWCUyIgdVK1slXVZaCCtZFAMWDg%3D%3D"
     }
 }